feat: Make plots function pure, delete old xpcs_plots.py file
It's a little ugly to have all that code in one function, but it
means we can use it in funcx without needing to register a container!
